refactor(frontend): consolidate article card variants into single component (#157)
Replace ArticleCardBase and three thin wrappers with a unified ArticleCard using variant props, plus CardTitle, CardMetadata, CardActions, and ArticleCardLayout sub-components.
